+++ app/src/main/AndroidManifest.xml
@@ -5,9 +5,6 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
 
-    <!-- Disable automatic updates in F-Droid -->
-    <!-- <uses-permission android:name="android.permission.REQUEST_INSTALL_PACKAGES" />-->
-
     <application
         android:name=".App"
         android:allowBackup="true"
