+++ app/src/main/java/me/ash/reader/data/model/Filter.kt
@@ -1,4 +1,4 @@
-package me.ash.reader.data.entity
+package me.ash.reader.data.model
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.FiberManualRecord
@@ -6,7 +6,10 @@ import androidx.compose.material.icons.rounded.FiberManualRecord
 import androidx.compose.material.icons.rounded.Star
 import androidx.compose.material.icons.rounded.StarOutline
 import androidx.compose.material.icons.rounded.Subject
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import me.ash.reader.R
 
 class Filter(
     val index: Int,
@@ -33,5 +36,13 @@ class Filter(
             iconOutline = Icons.Rounded.Subject,
             iconFilled = Icons.Rounded.Subject,
         )
+        val values = listOf(Starred, Unread, All)
     }
+}
+
+@Composable
+fun Filter.getName(): String = when (this) {
+    Filter.Unread -> stringResource(R.string.unread)
+    Filter.Starred -> stringResource(R.string.starred)
+    else -> stringResource(R.string.all)
 }
