+++ app/src/main/java/me/ash/reader/data/model/LatestRelease.kt
@@ -1,0 +1,2 @@
+package me.ash.reader.data.model
+
