+++ app/src/main/java/me/ash/reader/data/module/OkHttpClientModule.kt
@@ -1,3 +1,23 @@
+/*
+ * Feeder: Android RSS reader app
+ * https://gitlab.com/spacecowboy/Feeder
+ *
+ * Copyright (C) 2022  Jonas Kalderstam
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package me.ash.reader.data.module
 
 import android.content.Context
@@ -7,11 +27,20 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import me.ash.reader.BuildConfig
-import me.ash.reader.cachingHttpClient
+import okhttp3.Cache
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.Response
+import java.io.File
+import java.security.KeyManagementException
+import java.security.NoSuchAlgorithmException
+import java.security.cert.X509Certificate
+import java.util.concurrent.TimeUnit
 import javax.inject.Singleton
+import javax.net.ssl.HostnameVerifier
+import javax.net.ssl.SSLContext
+import javax.net.ssl.TrustManager
+import javax.net.ssl.X509TrustManager
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -26,6 +55,56 @@ object OkHttpClientModule {
     ).newBuilder()
         .addNetworkInterceptor(UserAgentInterceptor)
         .build()
+}
+
+fun cachingHttpClient(
+    cacheDirectory: File? = null,
+    cacheSize: Long = 10L * 1024L * 1024L,
+    trustAllCerts: Boolean = true,
+    connectTimeoutSecs: Long = 30L,
+    readTimeoutSecs: Long = 30L
+): OkHttpClient {
+    val builder: OkHttpClient.Builder = OkHttpClient.Builder()
+
+    if (cacheDirectory != null) {
+        builder.cache(Cache(cacheDirectory, cacheSize))
+    }
+
+    builder
+        .connectTimeout(connectTimeoutSecs, TimeUnit.SECONDS)
+        .readTimeout(readTimeoutSecs, TimeUnit.SECONDS)
+        .followRedirects(true)
+
+    if (trustAllCerts) {
+        builder.trustAllCerts()
+    }
+
+    return builder.build()
+}
+
+fun OkHttpClient.Builder.trustAllCerts() {
+    try {
+        val trustManager = object : X509TrustManager {
+            override fun checkClientTrusted(chain: Array<out X509Certificate>?, authType: String?) {
+            }
+
+            override fun checkServerTrusted(chain: Array<out X509Certificate>?, authType: String?) {
+            }
+
+            override fun getAcceptedIssuers(): Array<X509Certificate> = emptyArray()
+        }
+
+        val sslContext = SSLContext.getInstance("TLS")
+        sslContext.init(null, arrayOf<TrustManager>(trustManager), null)
+        val sslSocketFactory = sslContext.socketFactory
+
+        sslSocketFactory(sslSocketFactory, trustManager)
+            .hostnameVerifier(HostnameVerifier { _, _ -> true })
+    } catch (e: NoSuchAlgorithmException) {
+        // ignore
+    } catch (e: KeyManagementException) {
+        // ignore
+    }
 }
 
 object UserAgentInterceptor : Interceptor {
