+++ app/src/main/java/me/ash/reader/data/repository/AbstractRssRepository.kt
@@ -2,11 +2,11 @@ package me.ash.reader.data.repository
 
 import android.content.Context
 import android.util.Log
-import androidx.hilt.work.HiltWorker
 import androidx.paging.PagingSource
-import androidx.work.*
-import dagger.assisted.Assisted
-import dagger.assisted.AssistedInject
+import androidx.work.CoroutineWorker
+import androidx.work.ExistingPeriodicWorkPolicy
+import androidx.work.ListenableWorker
+import androidx.work.WorkManager
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOn
@@ -15,9 +15,9 @@ import me.ash.reader.data.dao.ArticleDao
 import me.ash.reader.data.dao.FeedDao
 import me.ash.reader.data.dao.GroupDao
 import me.ash.reader.data.entity.*
+import me.ash.reader.data.model.ImportantCount
 import me.ash.reader.ui.ext.currentAccountId
 import java.util.*
-import java.util.concurrent.TimeUnit
 
 abstract class AbstractRssRepository constructor(
     private val context: Context,
@@ -130,10 +130,6 @@ abstract class AbstractRssRepository constructor(
         return feedDao.queryByLink(context.currentAccountId, url).isNotEmpty()
     }
 
-    fun peekWork(): String {
-        return workManager.getWorkInfosByTag("sync").get().size.toString()
-    }
-
     suspend fun updateGroup(group: Group) {
         groupDao.update(group)
     }
@@ -205,36 +201,5 @@ abstract class AbstractRssRepository constructor(
                 else -> articleDao.searchArticleWhenAll(accountId, content)
             }
         }
-    }
-}
-
-@HiltWorker
-class SyncWorker @AssistedInject constructor(
-    @Assisted context: Context,
-    @Assisted workerParams: WorkerParameters,
-    private val rssRepository: RssRepository,
-) : CoroutineWorker(context, workerParams) {
-
-    override suspend fun doWork(): Result {
-        Log.i("RLog", "doWork: ")
-        return rssRepository.get().sync(this)
-    }
-
-    companion object {
-        const val WORK_NAME = "article.sync"
-
-        val UUID: UUID
-
-        val repeatingRequest = PeriodicWorkRequestBuilder<SyncWorker>(
-            15, TimeUnit.MINUTES
-        ).setConstraints(
-            Constraints.Builder()
-                .build()
-        ).addTag(WORK_NAME).build().also {
-            UUID = it.id
-        }
-
-        fun setIsSyncing(boolean: Boolean) = workDataOf("isSyncing" to boolean)
-        fun Data.getIsSyncing(): Boolean = getBoolean("isSyncing", false)
     }
 }
