+++ app/src/main/java/me/ash/reader/data/repository/LocalRssRepository.kt
@@ -1,12 +1,7 @@
 package me.ash.reader.data.repository
 
-import android.app.*
 import android.content.Context
-import android.content.Intent
-import android.graphics.BitmapFactory
 import android.util.Log
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
 import androidx.work.CoroutineWorker
 import androidx.work.ListenableWorker
 import androidx.work.WorkManager
@@ -15,8 +10,6 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.withContext
-import me.ash.reader.MainActivity
-import me.ash.reader.R
 import me.ash.reader.data.dao.AccountDao
 import me.ash.reader.data.dao.ArticleDao
 import me.ash.reader.data.dao.FeedDao
@@ -30,8 +23,6 @@ import me.ash.reader.data.module.DispatcherIO
 import me.ash.reader.data.repository.SyncWorker.Companion.setIsSyncing
 import me.ash.reader.ui.ext.currentAccountId
 import me.ash.reader.ui.ext.spacerDollar
-import me.ash.reader.ui.page.common.ExtraName
-import me.ash.reader.ui.page.common.NotificationGroupName
 import java.util.*
 import javax.inject.Inject
 
@@ -41,6 +32,7 @@ class LocalRssRepository @Inject constructor(
     private val articleDao: ArticleDao,
     private val feedDao: FeedDao,
     private val rssHelper: RssHelper,
+    private val notificationHelper: NotificationHelper,
     private val accountDao: AccountDao,
     private val groupDao: GroupDao,
     @DispatcherDefault
@@ -52,16 +44,6 @@ class LocalRssRepository @Inject constructor(
     context, accountDao, articleDao, groupDao,
     feedDao, workManager, dispatcherIO
 ) {
-    private val notificationManager: NotificationManagerCompat =
-        NotificationManagerCompat.from(context).apply {
-            createNotificationChannel(
-                NotificationChannel(
-                    NotificationGroupName.ARTICLE_UPDATE,
-                    NotificationGroupName.ARTICLE_UPDATE,
-                    NotificationManager.IMPORTANCE_DEFAULT
-                )
-            )
-        }
 
     override suspend fun updateArticleInfo(article: Article) {
         articleDao.update(article)
@@ -98,7 +80,7 @@ class LocalRssRepository @Inject constructor(
                 .awaitAll()
                 .forEach {
                     if (it.isNotify) {
-                        notify(
+                        notificationHelper.notify(
                             FeedWithArticle(
                                 it.feedWithArticle.feed,
                                 articleDao.insertIfNotExist(it.feedWithArticle.articles)
@@ -182,76 +164,5 @@ class LocalRssRepository @Inject constructor(
             feedWithArticle = FeedWithArticle(feed, articles),
             isNotify = articles.isNotEmpty() && feed.isNotification
         )
-    }
-
-    private fun notify(
-        feedWithArticle: FeedWithArticle,
-    ) {
-        notificationManager.createNotificationChannelGroup(
-            NotificationChannelGroup(
-                feedWithArticle.feed.id,
-                feedWithArticle.feed.name
-            )
-        )
-        feedWithArticle.articles.forEach { article ->
-            val builder = NotificationCompat.Builder(context, NotificationGroupName.ARTICLE_UPDATE)
-                .setSmallIcon(R.drawable.ic_notification)
-                .setLargeIcon(
-                    (BitmapFactory.decodeResource(
-                        context.resources,
-                        R.drawable.ic_notification
-                    ))
-                )
-                .setContentTitle(article.title)
-                .setContentIntent(
-                    PendingIntent.getActivity(
-                        context,
-                        Random().nextInt() + article.id.hashCode(),
-                        Intent(context, MainActivity::class.java).apply {
-                            flags = Intent.FLAG_ACTIVITY_NEW_TASK or
-                                    Intent.FLAG_ACTIVITY_CLEAR_TASK
-                            putExtra(
-                                ExtraName.ARTICLE_ID,
-                                article.id
-                            )
-                        },
-                        PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-                    )
-                )
-                .setGroup(feedWithArticle.feed.id)
-                .setStyle(
-                    NotificationCompat.BigTextStyle()
-                        .bigText(article.shortDescription)
-                        .setSummaryText(feedWithArticle.feed.name)
-                )
-
-            notificationManager.notify(
-                Random().nextInt() + article.id.hashCode(),
-                builder.build().apply {
-                    flags = Notification.FLAG_AUTO_CANCEL
-                }
-            )
-        }
-
-        if (feedWithArticle.articles.size > 1) {
-            notificationManager.notify(
-                Random().nextInt() + feedWithArticle.feed.id.hashCode(),
-                NotificationCompat.Builder(context, NotificationGroupName.ARTICLE_UPDATE)
-                    .setSmallIcon(R.drawable.ic_notification)
-                    .setLargeIcon(
-                        (BitmapFactory.decodeResource(
-                            context.resources,
-                            R.drawable.ic_notification
-                        ))
-                    )
-                    .setStyle(
-                        NotificationCompat.InboxStyle()
-                            .setSummaryText(feedWithArticle.feed.name)
-                    )
-                    .setGroup(feedWithArticle.feed.id)
-                    .setGroupSummary(true)
-                    .build()
-            )
-        }
     }
 }
