+++ app/src/main/java/me/ash/reader/data/repository/NotificationHelper.kt
@@ -1,0 +1,103 @@
+package me.ash.reader.data.repository
+
+import android.app.*
+import android.content.Context
+import android.content.Intent
+import android.graphics.BitmapFactory
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
+import dagger.hilt.android.qualifiers.ApplicationContext
+import me.ash.reader.MainActivity
+import me.ash.reader.R
+import me.ash.reader.data.entity.FeedWithArticle
+import me.ash.reader.ui.page.common.ExtraName
+import me.ash.reader.ui.page.common.NotificationGroupName
+import java.util.*
+import javax.inject.Inject
+
+class NotificationHelper @Inject constructor(
+    @ApplicationContext
+    private val context: Context,
+) {
+    private val notificationManager: NotificationManagerCompat =
+        NotificationManagerCompat.from(context).apply {
+            createNotificationChannel(
+                NotificationChannel(
+                    NotificationGroupName.ARTICLE_UPDATE,
+                    NotificationGroupName.ARTICLE_UPDATE,
+                    NotificationManager.IMPORTANCE_DEFAULT
+                )
+            )
+        }
+
+    fun notify(
+        feedWithArticle: FeedWithArticle,
+    ) {
+        notificationManager.createNotificationChannelGroup(
+            NotificationChannelGroup(
+                feedWithArticle.feed.id,
+                feedWithArticle.feed.name
+            )
+        )
+        feedWithArticle.articles.forEach { article ->
+            val builder = NotificationCompat.Builder(context, NotificationGroupName.ARTICLE_UPDATE)
+                .setSmallIcon(R.drawable.ic_notification)
+                .setLargeIcon(
+                    (BitmapFactory.decodeResource(
+                        context.resources,
+                        R.drawable.ic_notification
+                    ))
+                )
+                .setContentTitle(article.title)
+                .setContentIntent(
+                    PendingIntent.getActivity(
+                        context,
+                        Random().nextInt() + article.id.hashCode(),
+                        Intent(context, MainActivity::class.java).apply {
+                            flags = Intent.FLAG_ACTIVITY_NEW_TASK or
+                                    Intent.FLAG_ACTIVITY_CLEAR_TASK
+                            putExtra(
+                                ExtraName.ARTICLE_ID,
+                                article.id
+                            )
+                        },
+                        PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
+                    )
+                )
+                .setGroup(feedWithArticle.feed.id)
+                .setStyle(
+                    NotificationCompat.BigTextStyle()
+                        .bigText(article.shortDescription)
+                        .setSummaryText(feedWithArticle.feed.name)
+                )
+
+            notificationManager.notify(
+                Random().nextInt() + article.id.hashCode(),
+                builder.build().apply {
+                    flags = Notification.FLAG_AUTO_CANCEL
+                }
+            )
+        }
+
+        if (feedWithArticle.articles.size > 1) {
+            notificationManager.notify(
+                Random().nextInt() + feedWithArticle.feed.id.hashCode(),
+                NotificationCompat.Builder(context, NotificationGroupName.ARTICLE_UPDATE)
+                    .setSmallIcon(R.drawable.ic_notification)
+                    .setLargeIcon(
+                        (BitmapFactory.decodeResource(
+                            context.resources,
+                            R.drawable.ic_notification
+                        ))
+                    )
+                    .setStyle(
+                        NotificationCompat.InboxStyle()
+                            .setSummaryText(feedWithArticle.feed.name)
+                    )
+                    .setGroup(feedWithArticle.feed.id)
+                    .setGroupSummary(true)
+                    .build()
+            )
+        }
+    }
+}
