+++ app/src/main/java/me/ash/reader/ui/ext/DateExt.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.core.os.ConfigurationCompat
 import me.ash.reader.R
 import java.text.DateFormat
+import java.text.ParsePosition
 import java.text.SimpleDateFormat
 import java.util.*
 
@@ -40,4 +41,27 @@ fun Date.formatAsString(
             }
         }
     }
+}
+
+private fun String.parseToDate(
+    patterns: Array<String> = arrayOf(
+        "yyyy-MM-dd'T'HH:mm:ss'Z'",
+        "yyyy-MM-dd",
+        "yyyy-MM-dd HH:mm:ss",
+        "yyyyMMdd",
+        "yyyy/MM/dd",
+        "yyyy年MM月dd日",
+        "yyyy MM dd",
+    )
+): Date? {
+    val df = SimpleDateFormat()
+    for (pattern in patterns) {
+        df.applyPattern(pattern)
+        df.isLenient = false
+        val date = df.parse(this, ParsePosition(0))
+        if (date != null) {
+            return date
+        }
+    }
+    return null
 }
