+++ app/src/main/java/me/ash/reader/ui/page/home/feeds/FeedsPage.kt
@@ -28,7 +28,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import kotlinx.coroutines.flow.map
 import me.ash.reader.R
-import me.ash.reader.data.entity.toVersion
+import me.ash.reader.data.model.getName
+import me.ash.reader.data.model.toVersion
 import me.ash.reader.data.preference.*
 import me.ash.reader.data.repository.SyncWorker.Companion.getIsSyncing
 import me.ash.reader.ui.component.Banner
